+++ tests/fixtures/data/campaigns.php
@@ -8,4 +8,8 @@ return [
         'campaignTypeId' => $this->campaignTypeIds['campaignType1'] ?? null,
         'title' => 'Campaign 1',
     ],
+    [
+        'campaignTypeId' => $this->campaignTypeIds['campaignType2'] ?? null,
+        'title' => 'Campaign 2',
+    ],
 ];