+++ tests/fixtures/data/sendouts.php
@@ -4,6 +4,7 @@
  */
 
 use craft\helpers\Db;
+use putyourlightson\campaign\elements\CampaignElement;
 use putyourlightson\campaign\elements\SendoutElement;
 
 return [
@@ -23,6 +24,19 @@ return [
     [
         'title' => 'Sendout 2',
         'subject' => 'Subject 2',
+        'sendoutType' => 'regular',
+        'sendStatus' => SendoutElement::STATUS_PENDING,
+        'sendDate' => Db::prepareDateForDb(new DateTime()),
+        'campaignId' => CampaignElement::find()->campaignType('campaignType2')->one()->id,
+        'mailingListIds' => $this->mailingListIds,
+        'segmentIds' => $this->segmentIds,
+        'fromName' => 'From Name',
+        'fromEmail' => 'from@email.com',
+        'notificationEmailAddress' => 'notify@email.com',
+    ],
+    [
+        'title' => 'Sendout 3',
+        'subject' => 'Subject 3',
         'sendoutType' => 'automated',
         'sendStatus' => SendoutElement::STATUS_PENDING,
         'campaignId' => $this->campaignId,