+++ tests/unit/controllers/FormsControllerTest.php
@@ -3,7 +3,7 @@
  * @copyright Copyright (c) PutYourLightsOn
  */
 
-namespace putyourlightson\campaigntests\unit\controllers;
+namespace putyourlightson\campaigntests\unit\base;
 
 use putyourlightson\campaign\elements\ContactElement;
 use putyourlightson\campaign\elements\MailingListElement;